refactor: change tcp package https://gravitee.atlassian.net/browse/APIM-3224
