+++ gravitee-apim-gateway/gravitee-apim-gateway-reactor/src/main/java/io/gravitee/gateway/reactive/reactor/DefaultTcpSocketDispatcher.java
@@ -24,8 +24,8 @@ import io.gravitee.gateway.reactive.api.context.InternalContextAttributes;
 import io.gravitee.gateway.reactive.core.context.DefaultExecutionContext;
 import io.gravitee.gateway.reactive.reactor.handler.TcpAcceptorResolver;
 import io.gravitee.gateway.reactor.handler.TcpAcceptor;
-import io.gravitee.gateway.tcp.reactive.VertxTcpRequest;
-import io.gravitee.gateway.tcp.reactive.VertxTcpResponse;
+import io.gravitee.gateway.reactive.tcp.VertxTcpRequest;
+import io.gravitee.gateway.reactive.tcp.VertxTcpResponse;
 import io.reactivex.rxjava3.core.Completable;
 import io.vertx.rxjava3.core.net.NetSocket;
 import lombok.RequiredArgsConstructor;
