+++ gravitee-apim-gateway/gravitee-apim-gateway-tcp/src/main/java/io/gravitee/gateway/reactive/tcp/VertxTcpResponse.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.gravitee.gateway.tcp.reactive;
+package io.gravitee.gateway.reactive.tcp;
 
 import io.gravitee.gateway.api.http.HttpHeaders;
 import io.gravitee.gateway.reactive.api.context.GenericExecutionContext;
