+++ gravitee-apim-gateway/gravitee-apim-gateway-tcp/src/test/java/io/gravitee/gateway/reactive/tcp/VertxReadStreamUtilTest.java
@@ -13,12 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.gravitee.gateway.tcp.reactive;
+package io.gravitee.gateway.reactive.tcp;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import io.gravitee.gateway.api.buffer.Buffer;
-import io.gravitee.gateway.tcp.reactive.vertx.RecordingWriteStream;
+import io.gravitee.gateway.reactive.tcp.vertx.RecordingWriteStream;
 import io.reactivex.rxjava3.core.Flowable;
 import io.vertx.rxjava3.core.streams.WriteStream;
 import org.junit.jupiter.api.DisplayNameGeneration;
