+++ gravitee-apim-gateway/gravitee-apim-gateway-tcp/src/test/java/io/gravitee/gateway/reactive/tcp/VertxTcpResponseTest.java
@@ -13,9 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.gravitee.gateway.tcp.reactive;
+package io.gravitee.gateway.reactive.tcp;
 
-import static io.gravitee.gateway.tcp.reactive.VertxReadStreamUtil.toVertxRxReadStream;
+import static io.gravitee.gateway.reactive.tcp.VertxReadStreamUtil.toVertxRxReadStream;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.params.provider.Arguments.arguments;
 import static org.mockito.Mockito.when;
@@ -24,7 +24,7 @@ import io.gravitee.common.http.IdGenerator;
 import io.gravitee.common.utils.UUID;
 import io.gravitee.gateway.api.buffer.Buffer;
 import io.gravitee.gateway.reactive.api.context.ExecutionContext;
-import io.gravitee.gateway.tcp.reactive.vertx.RecordingWriteStream;
+import io.gravitee.gateway.reactive.tcp.vertx.RecordingWriteStream;
 import io.reactivex.rxjava3.core.Flowable;
 import io.vertx.rxjava3.core.net.NetSocket;
 import io.vertx.rxjava3.core.streams.WriteStream;
