+++ gravitee-apim-gateway/gravitee-apim-gateway-tcp/src/test/java/io/gravitee/gateway/reactive/tcp/vertx/RecordingWriteStreamTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.gravitee.gateway.tcp.reactive.vertx;
+package io.gravitee.gateway.reactive.tcp.vertx;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.fail;
