+++ gravitee-apim-plugin/gravitee-apim-plugin-endpoint/gravitee-apim-plugin-endpoint-tcp-proxy/src/test/java/io/gravitee/plugin/endpoint/tcp/proxy/TcpProxyEndpointConnectorTest.java
@@ -22,14 +22,11 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import io.gravitee.common.utils.UUID;
 import io.gravitee.gateway.core.component.CustomComponentProvider;
 import io.gravitee.gateway.reactive.api.context.DeploymentContext;
-import io.gravitee.gateway.reactive.api.context.InternalContextAttributes;
 import io.gravitee.gateway.reactive.api.helper.PluginConfigurationHelper;
 import io.gravitee.gateway.reactive.core.context.DefaultExecutionContext;
-import io.gravitee.gateway.tcp.reactive.VertxTcpRequest;
-import io.gravitee.gateway.tcp.reactive.VertxTcpResponse;
-import io.reactivex.rxjava3.core.Completable;
+import io.gravitee.gateway.reactive.tcp.VertxTcpRequest;
+import io.gravitee.gateway.reactive.tcp.VertxTcpResponse;
 import io.reactivex.rxjava3.core.Flowable;
-import io.vertx.junit5.VertxExtension;
 import io.vertx.rxjava3.core.Vertx;
 import io.vertx.rxjava3.core.net.NetServer;
 import io.vertx.rxjava3.core.net.NetSocket;
